Rename NotificationQueue to NotificationQueueRabbit; make construction dependent on RabbitMQ parameters #51
